# Convert regexp capture groups to non-capture groups