--- a/query/parse.go
+++ b/query/parse.go
@@ -231,6 +231,15 @@ func regexpQuery(text string, content, file bool) (Q, error) {
 		return nil, err
 	}
 
+	if hasCapture(r) {
+		re, err := ConvertCapture(r)
+		if err == nil {
+			r = re
+		} else {
+			log.Printf("WARN: ConvertCapture error: %s", err)
+		}
+	}
+
 	if r.Op == syntax.OpLiteral {
 		expr = &Substring{
 			Pattern:  string(r.Rune),