Progress on https://github.com/sourcegraph/sourcegraph/issues/34826

When a regexp query is parsed if it contains capture groups the regexp is rewritten to use non-capture groups.
In addition to the speed benefit described in the above issue this also allows certain regexp to be rewritten as Substring queries. e.g. The Regexp query `ba(na)na` becomes `ba(?:na)na` which can be simplified to the literal `banana` and used as a Substring query.

### Test Plan
Tested locally:
- Added `TestConvertCapture` to verify correct behaviour converting from capture to non-capture group
- Updated `TestParseQuery` cases where a Regexp query is now converted to a Substring query
- Benchmarks `^BenchmarkShardedSearch$/regexp` comparing performance of parsing the same query but using capture and non-capture groups to measure the overhead of converting the capture group case to non-capture
Here is a comparison of running the benchmark without (old time) and with (new time) the optimisation
```
name                                             old time/op    new time/op    delta
ShardedSearch/regexp_all_results_capture-16         1.26s ±14%     0.96s ± 7%  -23.83%  (p=0.008 n=5+5)
ShardedSearch/regexp_all_results_non-capture-16     934ms ± 5%     941ms ± 1%     ~     (p=0.690 n=5+5)

name                                             old alloc/op   new alloc/op   delta
ShardedSearch/regexp_all_results_capture-16        3.77GB ± 0%    3.01GB ± 0%  -20.30%  (p=0.008 n=5+5)
ShardedSearch/regexp_all_results_non-capture-16    3.01GB ± 0%    3.01GB ± 0%     ~     (p=1.000 n=5+5)

name                                             old allocs/op  new allocs/op  delta
ShardedSearch/regexp_all_results_capture-16         33.7M ± 0%     26.2M ± 0%  -22.14%  (p=0.016 n=5+4)
ShardedSearch/regexp_all_results_non-capture-16     26.2M ± 0%     26.2M ± 0%     ~     (p=0.857 n=5+5)
```

Additionally I ran `zoekt-webserver` verifying correct behaviour with different regexp queries trying to detect any edge cases. As a safeguard the code modifies a copy of the original regexp such that in the extremely unlikely event of an edge case which results in an error the code can fallback and use the original regexp query